Changed Plot, Action and User ids to be longs instead of ints

+++ RealFarm/src/com/commonsensenet/realfarm/Global.java
@@ -7,10 +7,10 @@ public class Global {
 	public static boolean cameraFlag;
 	public static String deviceId = null;
 	public static boolean isAudioEnabled = true;
-	public static int plotId = -1;
+	public static long plotId = -1;
 	public static String plotImagePath;
 	public static Bitmap rotated;
 	public static Class<?> selectedAction = null;
 	public static int selectedVideo;
-	public static int userId = -1;
+	public static long userId = -1;
 }
